Move the model implementations out of the recipes dir for now
