+++ tests/models/jax/test_llama4.py
@@ -10,8 +10,7 @@ from flax.typing import PRNGKey
 from jax.sharding import Mesh
 from vllm.config import ModelConfig
 
-from tpu_commons.models.jax.recipes.llama4 import (Llama4ForCausalLM,
-                                                   Llama4WeightLoader)
+from tpu_commons.models.jax.llama4 import Llama4ForCausalLM, Llama4WeightLoader
 
 
 class MockParamLlama4:
@@ -112,7 +111,7 @@ class TestLlama4ForCausalLM:
 
         assert jnp.all(final_norm_scale == 1.0)
 
-    @patch("tpu_commons.models.jax.recipes.llama4.Llama4WeightLoader")
+    @patch("tpu_commons.models.jax.llama4.Llama4WeightLoader")
     def test_load_weights_called_correctly(self, mock_loader_cls, rng, mesh):
         """Tests that the weight loader is called correctly for checkpoint loading."""
         vllm_config = MockVllmConfig(model_name="llama4-scout",
@@ -177,8 +176,8 @@ class TestLlama4WeightLoader:
         # Mock get_param to return a mock param with the target shape (vocab_size, hidden_size)
         mock_param = MockParamLlama4(shape=(128, 32))
 
-        with patch("tpu_commons.models.jax.recipes.llama4.get_param", return_value=mock_param), \
-            patch("tpu_commons.models.jax.recipes.llama4.shard_put", return_value=jnp.ones(mock_param.value.shape)) as mock_shard_put:
+        with patch("tpu_commons.models.jax.llama4.get_param", return_value=mock_param), \
+            patch("tpu_commons.models.jax.llama4.shard_put", return_value=jnp.ones(mock_param.value.shape)) as mock_shard_put:
 
             # This will now pass after the code fix
             weight_loader.load_weights(model)
@@ -217,8 +216,8 @@ class TestLlama4WeightLoader:
              dummy_weight),
         ]
 
-        with patch("tpu_commons.models.jax.recipes.llama4.get_param", return_value=mock_param), \
-            patch("tpu_commons.models.jax.recipes.llama4.shard_put", return_value=jnp.ones(mock_param.value.shape)) as mock_shard_put:
+        with patch("tpu_commons.models.jax.llama4.get_param", return_value=mock_param), \
+            patch("tpu_commons.models.jax.llama4.shard_put", return_value=jnp.ones(mock_param.value.shape)) as mock_shard_put:
 
             # Call _map_llama4_gate_up_proj directly
             weight_loader._map_llama4_gate_up_proj(
