+++ tpu_commons/models/jax/model_loader.py
@@ -73,8 +73,8 @@ def _get_model_architecture(config: PretrainedConfig) -> nnx.Module:
     if os.getenv("NEW_MODEL_DESIGN", False):
         from tpu_commons.experimental.llama3_jax_stashed import \
             LlamaForCausalLM
-        from tpu_commons.models.jax.recipes.deepseek_v3 import DeepSeekV3
-        from tpu_commons.models.jax.recipes.llama4 import Llama4ForCausalLM
+        from tpu_commons.models.jax.deepseek_v3 import DeepSeekV3
+        from tpu_commons.models.jax.llama4 import Llama4ForCausalLM
         _MODEL_REGISTRY["DeepSeekV3"] = DeepSeekV3
         _MODEL_REGISTRY["LlamaForCausalLM"] = LlamaForCausalLM
         _MODEL_REGISTRY["Llama4ForCausalLM"] = Llama4ForCausalLM
